+++ Content.Client/Bank/BUI/WithdrawlBankATMMenuBoundUserInterface.cs
@@ -32,7 +32,7 @@ public sealed class WithdrawBankATMMenuBoundUserInterface : BoundUserInterface
 
     private void OnWithdraw()
     {
-        if (_menu?.Amount is not int amount)
+        if (_menu?.Amount is not ulong amount)
             return;
 
         SendMessage(new BankWithdrawMessage(amount));
